+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add support for `?` operator in functions and closures
+
 # 0.3.8 - 2019-01-26
 
 * Update minimum derive_utils version to 0.6.1